Refactor hdl to aurrigo
