--- docker/noetic_llvm/Dockerfile
@@ -1,25 +1,0 @@
-FROM ros:noetic
-
-RUN apt-get update && apt-get install --no-install-recommends -y \
-    && apt-get install --no-install-recommends -y wget nano build-essential \
-    git clang lld libomp-dev libgtest-dev libboost-all-dev libopencv-dev ros-noetic-pcl-ros \
-    && apt-get clean \
-    && rm -rf /var/lib/apt/lists/*
-
-RUN update-alternatives --install /usr/bin/ld ld /usr/bin/ld.lld 50
-
-RUN mkdir -p /root/catkin_ws/src
-WORKDIR /root/catkin_ws/src
-RUN /bin/bash -c '. /opt/ros/noetic/setup.bash; catkin_init_workspace'
-
-COPY . /root/catkin_ws/src/hdl_global_localization/
-WORKDIR /root/catkin_ws/src/hdl_global_localization
-
-WORKDIR /root/catkin_ws
-RUN /bin/bash -c '. /opt/ros/noetic/setup.bash; CC=clang CXX=clang++ catkin_make -DENABLE_TEASER=OFF'
-RUN sed -i "6i source \"/root/catkin_ws/devel/setup.bash\"" /ros_entrypoint.sh
-
-WORKDIR /
-
-ENTRYPOINT ["/ros_entrypoint.sh"]
-CMD ["bash"]
