+++ package.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0"?>
 <package format="2">
-  <name>hdl_global_localization</name>
+  <name>aurrigo_global_localization</name>
   <version>0.0.0</version>
-  <description>The hdl_global_localization package</description>
+  <description>The aurrigo_global_localization package</description>
 
   <maintainer email="k.koide@aist.go.jp">k.koide</maintainer>
